In PR #2 I tried switching to PHP 5.6 and failed miserably. Apt wasn’t properly upgrading, Percona was bombing during provisioning, etc.

Upgrading some of the core Puppet deps and explicitly setting some requires around PPAs seem to have solved those issues.

So Pull this branch, then `vagrant reload --provision`. Please report if that works!

If it doesn’t, `vagrant destroy -f && vagrant up && vagrant reload --provison`. Please report if you need to do that instead.