@jeffgeorge 2016-01-19
Perf Change: Negligible. :weary:

Test Case: Run `$ ab -n 100 -c 10 http://sc.vm/` twice. First time to warm up OpCache, second time to record the speed.

Current Master:
```
Document Path:          /
Document Length:        48014 bytes

Concurrency Level:      10
Time taken for tests:   33.926 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      4850300 bytes
HTML transferred:       4801400 bytes
Requests per second:    2.95 [#/sec] (mean)
Time per request:       3392.558 [ms] (mean)
Time per request:       339.256 [ms] (mean, across all concurrent requests)
Transfer rate:          139.62 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:  1649 3344 742.9   3232    5387
Waiting:     1609 3228 721.2   3127    5294
Total:       1650 3345 742.9   3232    5387

Percentage of the requests served within a certain time (ms)
  50%   3232
  66%   3532
  75%   3611
  80%   3718
  90%   4473
  95%   4921
  98%   5374
  99%   5387
 100%   5387 (longest request)
```

This Branch:
```
Document Path:          /
Document Length:        48014 bytes

Concurrency Level:      10
Time taken for tests:   33.031 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      4851600 bytes
HTML transferred:       4801400 bytes
Requests per second:    3.03 [#/sec] (mean)
Time per request:       3303.096 [ms] (mean)
Time per request:       330.310 [ms] (mean, across all concurrent requests)
Transfer rate:          143.44 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:  1130 3239 523.7   3292    4132
Waiting:     1111 3133 506.0   3172    3994
Total:       1130 3239 523.7   3292    4132

Percentage of the requests served within a certain time (ms)
  50%   3292
  66%   3516
  75%   3625
  80%   3634
  90%   3734
  95%   3858
  98%   4041
  99%   4132
 100%   4132 (longest request)
```

It's better across the board, but Apache Bench tests are variable enough that I'd consider the difference within the margin of error.