@jeffgeorge 2016-01-19
From testing, it looks like you can't just checkout this branch and `vagrant reload --provision`, that'll add Ondrej's PPA but won't upgrade PHP. You can run `vagrant ssh -c "sudo apt-get upgrade php5"`, but that  upgrades . . other stuff. So upon switching to this branch I recommend destroying and rebuilding.