@jeffgeorge 2016-01-28
Merging after testing and internal favorable feedback. :shipit: